Make `core_impls` test target multi-file.
