+++ tests/core_impls_and_misc/main.rs
@@ -1,0 +1,5 @@
+#[path = "../helper/mod.rs"]
+mod helper;
+
+mod core_impls;
+mod iter;
